+++ src/apptk/view/view_component_base.py
@@ -8,6 +8,15 @@ from abc import ABC, abstractmethod
 
 
 class ViewComponentBase(ABC):
+    """A node in the view composite structure.
+    
+    Sub-view instances of the same class can be registered and unregistered.
+    
+    Passes down the following commands to the sub-views:
+        - refresh
+        - lock/unlock
+        - emable/disable menu    
+    """
 
     @abstractmethod
     def __init__(self, model, view, controller):
@@ -15,23 +24,49 @@ class ViewComponentBase(ABC):
         self._ui = view
         self._ctrl = controller
 
+        self._viewComponents = []
+        # applying the Composite design pattern
+
     def disable_menu(self):
         """Disable UI widgets, e.g. when no project is open."""
-        pass
+        for viewComponent in self._viewComponents:
+            viewComponent.disable_menu()
 
     def enable_menu(self):
         """Enable UI widgets, e.g. when a project is opened."""
-        pass
+        for viewComponent in self._viewComponents:
+            viewComponent.enable_menu()
 
     def lock(self):
         """Inhibit changes on the model."""
-        pass
+        for viewComponent in self._viewComponents:
+            viewComponent.lock()
 
     def refresh(self):
-        """Refresh the view after model change."""
-        pass
+        """Refresh all view components."""
+        for viewComponent in self._viewComponents:
+            viewComponent.refresh()
+
+    def register_view(self, viewComponent):
+        """Add a view object to the composite list.
+        
+        Positional arguments:
+            viewComponent -- Reference to a ViewComponentBase subclass instance.
+        """
+        if not viewComponent in self._viewComponents:
+            self._viewComponents.append(viewComponent)
 
     def unlock(self):
         """Enable changes on the model."""
-        pass
+        for viewComponent in self._viewComponents:
+            viewComponent.unlock()
+
+    def unregister_view(self, viewComponent):
+        """Revove a view object from the component list.
+        
+        Positional arguments:
+            viewComponent -- Reference to a ViewComponentBase subclass instance.
+        """
+        if viewComponent in self._viewComponents:
+            self._viewComponents.remove(viewComponent)